// src/lib/api/types.ts
export type ApiErrorResponse = {
  timestamp: string;
  status: number;
  error: string;
  message: string;
  path: string;
  requestId: string | null;
  fieldViolations: {
    field: string;
    message: string;
  }[];
};

export type AuthUser = {
  id: string;
  email: string;
  fullName: string;
};

export type RegisterResponse = {
  userId: string;
  organizationId: string;
  email: string;
  fullName: string;
  organizationName: string;
  role: "OWNER";
};

export type LoginResponse = {
  accessToken: string;
  refreshToken: string;
  tokenType: "Bearer";
  expiresInSeconds: number;
  user: AuthUser;
};